add View App for merit order use market IP in templates to start the right python scripts

+++ model/main.py
@@ -66,14 +66,16 @@ def meritOrder():
 
 @app.route('/meritOrder/plot', methods = ['GET'])
 @cross_origin()
-def plotMeritOrderHourly():
-
+def plotMeritOrder():
     seconds = int(request.args.get('from'))
     date = pd.to_datetime(seconds)
     hour = request.args.get('hour')
-    plot = getMeritOrder(mongoCon, date.date(), request.args.get('hour'))
+    try:
+        plot = getMeritOrder(mongoCon, date.date(), request.args.get('hour'))
+        return render_template('plotMeritOrder.html', **locals())
+    except:
+        return 'Page not found - Error 404'
 
-    return render_template('plotMeritOrder.html', **locals())
 
 # ----- Start Simulation -----
 @app.route('/run', methods=['POST'])
@@ -141,7 +143,6 @@ def simulation(start, end, valid=True):
             print('Error in Actual ' + str(date.date()))
             print(e)
 
-
 if __name__ == "__main__":
     pids = []
     # ----- InfluxDB -----
@@ -159,13 +160,14 @@ if __name__ == "__main__":
 
     tm.sleep(2)
 
-    try:
-        influxCon.influx.drop_database(database)
-    except:
-        pass
-    influxCon.influx.create_database(database)
-    for name in mongoCon.orderDB.list_collection_names():
-        mongoCon.orderDB.drop_collection(name)
+    if config.getboolean('Results', 'Delete'):
+        try:
+            influxCon.influx.drop_database(database)
+        except:
+            pass
+        influxCon.influx.create_database(database)
+        for name in mongoCon.orderDB.list_collection_names():
+            mongoCon.orderDB.drop_collection(name)
 
     init = mongoCon.orderDB["init"]
     query = {"_id": 'start'}
@@ -173,7 +175,10 @@ if __name__ == "__main__":
     init.update_one(filter=query, update=start, upsert=True)
 
     try:
-        app.run(debug=False, port=5010, host='0.0.0.0')
+        if config.getboolean('Market','Local'):
+            app.run(debug=False, port=config.getint('Market','Port'), host='127.0.0.1')
+        else:
+            app.run(debug=False, port=config.getint('Market','Port'), host='0.0.0.0')
     except Exception as e:
         print(e)
     finally:
